requirements.txt
streamlit
pandas
numpy
plotly
fpdf
psycopg2-binary
